exception handling added
* scripts will now inform you if there are no student matching the name input or if there is more than one math
* searching for students is now being done by a separate function

+++ scripts.py
@@ -3,19 +3,36 @@ from datacenter.models import Mark
 from datacenter.models import Chastisement
 from datacenter.models import Lesson
 from datacenter.models import Commendation
+from django.core.exceptions import ObjectDoesNotExist
+from django.core.exceptions import MultipleObjectsReturned
 import random
 
+def find_schoolkid(name):
+    try:
+        kid = Schoolkid.objects.get(full_name__contains=name)
+    except ObjectDoesNotExist as e:
+        return e
+    except MultipleObjectsReturned as e:
+        return e
+    return kid 
+
 def fix_marks(name):
-    kid = Schoolkid.objects.get(full_name__contains=name)
-    bad_marks = Mark.objects.filter(schoolkid=kid, points__lte=3)
-    for mark in bad_marks:
-        mark.points = 5
-        mark.save()
+    kid = find_schoolkid(name)
+    if type(kid) == Schoolkid:
+        bad_marks = Mark.objects.filter(schoolkid=kid, points__lte=3)
+        for mark in bad_marks:
+            mark.points = 5
+            mark.save()
+    else:
+        print(kid)
 
 def remove_chastisements(name):
-    kid = Schoolkid.objects.get(full_name__contains=name)
-    chastisements = Chastisement.objects.filter(schoolkid=kid)
-    chastisements.delete()
+    kid = find_schoolkid(name)
+    if type(kid) == Schoolkid:
+        chastisements = Chastisement.objects.filter(schoolkid=kid)
+        chastisements.delete()
+    else:
+        print(kid)
 
 def create_commendation(name, subject):
     commendations = [
@@ -50,10 +67,13 @@ def create_commendation(name, subject):
         'Ты многое сделал, я это вижу!',
         'Теперь у тебя точно все получится!',
     ]
-    kid = Schoolkid.objects.get(full_name__contains=name)
-    lessons = Lesson.objects.filter(year_of_study=kid.year_of_study, group_letter=kid.group_letter, subject__title=subject)
-    lessons = lessons.order_by('-date')
-    latest_lesson = lessons[0]
-    commendation = Commendation.objects.create(text=random.choice(commendations), created=latest_lesson.date, schoolkid=kid,
-        subject=latest_lesson.subject, teacher=latest_lesson.teacher)
-    commendation.save()
+    kid = find_schoolkid(name)
+    if type(kid) == Schoolkid:
+        lessons = Lesson.objects.filter(year_of_study=kid.year_of_study, group_letter=kid.group_letter, subject__title=subject)
+        lessons = lessons.order_by('-date')
+        latest_lesson = lessons[0]
+        commendation = Commendation.objects.create(text=random.choice(commendations), created=latest_lesson.date, schoolkid=kid,
+            subject=latest_lesson.subject, teacher=latest_lesson.teacher)
+        commendation.save()
+    else:
+        print(kid)
